requirements.txt
 asgiref==3.8.1
 certifi==2025.6.15
 charset-normalizer==3.4.2
 dj-database-url==3.0.0
 django==5.2.3
 django-bootstrap5==25.2
 django-filter==25.1
 djangorestframework==3.16.1
 djangorestframework-simplejwt==5.5.1
 gunicorn==23.0.0
 idna==3.10
 jmespath==1.0.1
 packaging==25.0
 pillow==11.2.1
 jmespath==1.0.1
 packaging==25.0
 pillow==11.2.1
 packaging==25.0
 pillow==11.2.1
 pillow==11.2.1
 psycopg2-binary==2.9.10
 pyjwt==2.10.1
 pymysql==1.1.2
 pymysql==1.1.2
 python-decouple==3.8
 requests==2.32.4
 sqlparse==0.5.3
 stripe==12.2.0
 typing-extensions==4.14.0
 tzdata==2025.2
 urllib3==2.5.0
 django-mathfilters==1.0.0